Moving more classes over to crud.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@3782 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Forms/FormPatientForms.cs
@@ -119,7 +119,7 @@ namespace OpenDental {
 				SheetUtil.CalculateHeights(sheet,this.CreateGraphics());
 				if(FormS.TerminalSend) {
 					sheet.InternalNote="";//because null not ok
-					sheet.ShowInTerminal=Sheets.GetBiggestShowInTerminal(PatNum)+1;
+					sheet.ShowInTerminal=(byte)(Sheets.GetBiggestShowInTerminal(PatNum)+1);
 					Sheets.SaveNewSheet(sheet);//save each sheet.
 				}
 			}
